ballmouse: add 'X' button support
I don't know why it used to work in kodi but now it doesn't. Map it to BTN_SIDE.
Note that triggers/shoulders/buttons in the middle of the controller aren't mapped
either. Use evtest to see. You can use evtest on the unmapped controller to see
what are the original keys being sent.